requirements.txt
aiohttp
aiosqlite
ttkthemes
reportlab
python-dotenv
maigret
holehe
pyinstaller